requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
better-exceptions==0.3.3
python-dotenv==1.0.0
httpx==0.25.0
pydantic==2.5.0
openai==1.3.5
google-generativeai==0.4.1